tbor/part_final tests: run the pre-chain-walk rejects on hardware
`PartFinal` had a blanket `#![cfg(feature = "emu")]`, so the command had
zero hardware coverage even though two of its gates never touch the
out-of-band cert transport.

Split the module by what each test needs, mirroring `part_init`:

* `fw_rejects` — the state gate and the policy-hash gate. Both pass an
  empty `certs` slice, so nothing is carried out of band and no SGL Data
  Block is referenced. Un-gated, so they run on emu **and** hardware.
* `chain_path` — everything that supplies a real PTA chain. Those
  certificates travel out of band, which only `ddi/emu` implements
  (`ddi/nix` rejects non-empty `oob_items`), so the submodule keeps the
  emu gate until the driver grows an OOB path.

`part_final_reject_policy_mismatch` now binds `known_good_part_policy()`
instead of building a POTA-specific policy: the policy-hash gate does not
care which trust anchor is present, and the canonical fixture already
populates POTA. That drops the test's dependency on the emu-only
`x509_fixture`, so no gate had to be relaxed elsewhere.

Hardware coverage for `PartFinal` goes from 0 to 2 tests; emu keeps all 8.

Copilot-Session: b6e9403d-692c-4903-97ee-87f2606151c3
